pyproject.toml
[tool.poetry]
name = "blat-tutorial"
version = "0.1.0"
description = ""
authors = ["Arnau Pont <arnau.pont@blat.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.8.2"
requests = "^2.32.3"
strenum = "^0.4.15"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
